requirements.txt
Flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
google-genai==0.3.0
python-dotenv==1.0.0
